Refactor layout and styles
These changes do not consider responsive design. They are improvements
to make the layout and styles more readable.

The page header was changed to separate the nav from the info about
the newspaper issue. So when more links are added in other pages
they can be styled correctly by the css.

More classes with slightly better names were added to be targeted later by
styles or javascript code.

Css code was updated with the new class names and was cleaned to
improve readibility to some extent. Should be improved later
when responsive design is taken into account or as more pages are
created for the site.